code-for-pelican-deployment/requirements.txt
anyio==4.12.1
blessings==1.7
blinker==1.9.0
docutils==0.22.4
feedgenerator==2.2.1
idna==3.11
Jinja2==3.1.6
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
pdf2image==1.17.0
pelican==4.11.0.post0
pillow==12.1.0
Pygments==2.18.0
pypdf==6.6.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
rich==14.2.0
six==1.17.0
typing_extensions==4.15.0
Unidecode==1.4.0
watchfiles==1.1.1
